+++ universe/ConditionAll.h
@@ -18,7 +18,7 @@ struct FO_COMMON_API All final : public Condition {
     void Eval(const ScriptingContext& parent_context, ObjectSet& matches,
               ObjectSet& non_matches, SearchDomain search_domain = SearchDomain::NON_MATCHES) const override;
     std::string Description(bool negated = false) const override;
-    std::string Dump(unsigned short ntabs = 0) const override;
+    std::string Dump(uint8_t ntabs = 0) const override;
     void SetTopLevelContent(const std::string& content_name) override
     {}
     unsigned int GetCheckSum() const override;
